requirements.txt
torch==1.13.1
torchvision==0.14.1
torchsummary==1.5.1
pytorch-lightning==1.9.3
numpy==1.24.2
pandas==1.5.2
matplotlib==3.7.0
opencv-python-headless==4.7.0.72
Pillow==9.4.0
seaborn==0.12.2
scikit-learn==1.2.1
seaborn==0.12.2
tifffile==2023.2.27
aicsimageio==4.9.4
imagecodecs==2023.1.23
dataclasses-json==0.5.7
PyYAML==6.0
psutil==5.9.0